Add all notebook tests
